Create first port and adjust types
