+++ lib/types/payment_methods_response.ex
@@ -2,42 +2,17 @@ defmodule BoldApi.Types.PaymentMethodsResponse do
   @moduledoc """
   PaymentMethods response struct.
   """
-  alias BoldApi.Types.Tax
 
-  @type status :: :active | :processing | :paid | :rejected | :cancelled | :expired
-
+  @type payment_method :: {String.t(), %{min: integer(), max: integer()}}
+  @type payload :: %{
+          payment_methods: list(payment_method())
+        }
+  @type errors :: list(String.t())
   @type t :: %__MODULE__{
-          api_version: String.t(),
-          id: String.t(),
-          total: float(),
-          subtotal: float(),
-          tip_amount: float(),
-          taxes: list(Tax.t()),
-          status: status(),
-          expiration_date: integer(),
-          creation_date: integer(),
-          description: String.t(),
-          payment_method: String.t(),
-          transaction_id: String.t(),
-          amount_type: :open | :closed,
-          is_sandbox: boolean()
+          payload: payload(),
+          errors: errors()
         }
 
   @enforce_keys []
-  defstruct [
-    :api_version,
-    :id,
-    :total,
-    :subtotal,
-    :tip_amount,
-    :taxes,
-    :status,
-    :expiration_date,
-    :creation_date,
-    :description,
-    :payment_method,
-    :transaction_id,
-    :amount_type,
-    :is_sandbox
-  ]
+  defstruct [:payload, :errors]
 end
